Update unit test to use CommonsChunkPlugin
