--- test/cases/compile/require-context/index.js
@@ -1,8 +1,0 @@
-it("should maintain require context", function() {
-	var context = {foo: "bar"};
-	(function(){
-		require(["./a", "./b"], function(a, b) {
-			this.foo.should.eql("bar");
-		}.bind(this));
-	}).call(context);
-});
