+++ test/configCases/async-commons-chunk/require-context/a.js
@@ -1,0 +1,1 @@
+module.exports = "a";
