+++ test/configCases/async-commons-chunk/require-context/b.js
@@ -1,0 +1,1 @@
+module.exports = "b";
