+++ test/configCases/async-commons-chunk/require-context/webpack.config.js
@@ -1,0 +1,9 @@
+var webpack = require("../../../../");
+
+module.exports = {
+	plugins: [
+		new webpack.optimize.CommonsChunkPlugin({
+			async: true
+		})
+	]
+};
